+++ Sources/Loader.swift
@@ -52,7 +52,7 @@ public class FileSystemLoader: Loader, CustomStringConvertible {
       }
 
       let content: String = try templatePath.read()
-      return try environment.templateClass.init(templateString: content, environment: environment, name: name)
+      return environment.templateClass.init(templateString: content, environment: environment, name: name)
     }
 
     throw TemplateDoesNotExist(templateNames: [name], loader: self)
@@ -65,8 +65,35 @@ public class FileSystemLoader: Loader, CustomStringConvertible {
 
         if templatePath.exists {
           let content: String = try templatePath.read()
-          return try environment.templateClass.init(templateString: content, environment: environment, name: templateName)
+          return environment.templateClass.init(templateString: content, environment: environment, name: templateName)
         }
+      }
+    }
+
+    throw TemplateDoesNotExist(templateNames: names, loader: self)
+  }
+}
+
+
+public class DictionaryLoader: Loader {
+  public let templates: [String: String]
+
+  public init(templates: [String: String]) {
+    self.templates = templates
+  }
+
+  public func loadTemplate(name: String, environment: Environment) throws -> Template {
+    if let content = templates[name] {
+      return environment.templateClass.init(templateString: content, environment: environment, name: name)
+    }
+
+    throw TemplateDoesNotExist(templateNames: [name], loader: self)
+  }
+
+  public func loadTemplate(names: [String], environment: Environment) throws -> Template {
+    for name in names {
+      if let content = templates[name] {
+        return environment.templateClass.init(templateString: content, environment: environment, name: name)
       }
     }